Refactor filename
