+++ sherpa/cpp_api/websocket/offline-websocket-server-impl.h
@@ -30,7 +30,7 @@
 #include "sherpa/cpp_api/offline-recognizer.h"
 #include "sherpa/cpp_api/parse-options.h"
 #include "sherpa/cpp_api/websocket/http-server.h"
-#include "sherpa/cpp_api/websocket/tee_stream.h"
+#include "sherpa/cpp_api/websocket/tee-stream.h"
 #include "websocketpp/config/asio_no_tls.hpp"  // TODO(fangjun): support TLS
 #include "websocketpp/server.hpp"
