+++ sherpa/cpp_api/websocket/offline-websocket-server.cc
@@ -16,8 +16,11 @@
  * limitations under the License.
  */
 
+#include <thread>  // NOLINT
+
 #include "asio.hpp"
-#include "sherpa/cpp_api/websocket/online_websocket_server_impl.h"
+#include "sherpa/cpp_api/parse-options.h"
+#include "sherpa/cpp_api/websocket/offline-websocket-server-impl.h"
 #include "sherpa/csrc/log.h"
 #include "torch/all.h"
 
@@ -26,15 +29,18 @@ Automatic speech recognition with sherpa using websocket.
 
 Usage:
 
-./bin/online_websocket_server --help
+./bin/offline_websocket_server --help
 
-./bin/online_websocket_server \
+./bin/offline_websocket_server \
   --use-gpu=false \
   --port=6006 \
+  --num-io-threads=3 \
   --num-work-threads=5 \
+  --max-batch-size=5 \
   --nn-model=/path/to/cpu.jit \
   --tokens=/path/to/tokens.txt \
   --decoding-method=greedy_search \
+  --doc-root=../sherpa/bin/web \
   --log-file=./log.txt
 )";
 
@@ -49,13 +55,14 @@ int32_t main(int32_t argc, char *argv[]) {
 
   sherpa::ParseOptions po(kUsageMessage);
 
-  sherpa::OnlineWebsocketServerConfig config;
+  sherpa::OfflineWebsocketServerConfig config;
+  sherpa::OfflineWebsocketDecoderConfig decoder_config;
 
   // the server will listen on this port, for both websocket and http
   int32_t port = 6006;
 
   // size of the thread pool for handling network connections
-  int32_t num_io_threads = 1;
+  int32_t num_io_threads = 3;
 
   // size of the thread pool for neural network computation and decoding
   int32_t num_work_threads = 5;
@@ -70,6 +77,7 @@ int32_t main(int32_t argc, char *argv[]) {
   po.Register("port", &port, "The port on which the server will listen.");
 
   config.Register(&po);
+  decoder_config.Register(&po);
 
   if (argc == 1) {
     po.PrintUsage();
@@ -85,15 +93,17 @@ int32_t main(int32_t argc, char *argv[]) {
   }
 
   config.Validate();
+  decoder_config.Validate();
 
   asio::io_context io_conn;  // for network connections
   asio::io_context io_work;  // for neural network and decoding
 
-  sherpa::OnlineWebsocketServer server(io_conn, io_work, config);
+  sherpa::OfflineWebsocketServer server(io_conn, io_work, config,
+                                        decoder_config);
   server.Run(port);
 
   SHERPA_LOG(INFO) << "Listening on: " << port << "\n";
-  // SHERPA_LOG(INFO) << "Number of I/O threads: " << num_io_threads << "\n";
+  SHERPA_LOG(INFO) << "Number of I/O threads: " << num_io_threads << "\n";
   SHERPA_LOG(INFO) << "Number of work threads: " << num_work_threads << "\n";
 
   // give some work to do for the io_work pool
@@ -115,9 +125,8 @@ int32_t main(int32_t argc, char *argv[]) {
   std::ostringstream os;
   os << "\nPlease access the HTTP server using the following address: \n\n";
   os << "http://localhost:" << port << "\n";
-#if 0
-  // TODO(fangjun): Enable it for HTTPS
   os << "http://127.0.0.1:" << port << "\n";
+
   asio::ip::tcp::resolver resolver(io_conn);
   auto iter = resolver.resolve(asio::ip::host_name(), "");
   asio::ip::tcp::resolver::iterator end;
@@ -129,7 +138,6 @@ int32_t main(int32_t argc, char *argv[]) {
       std::cout << "Error message: " << ec << "\n";
     }
   }
-#endif
   SHERPA_LOG(INFO) << os.str();
 
   io_conn.run();
