+++ sherpa/cpp_api/websocket/online-websocket-server-impl.h
@@ -33,7 +33,7 @@
 #include "sherpa/cpp_api/online_stream.h"
 #include "sherpa/cpp_api/parse-options.h"
 #include "sherpa/cpp_api/websocket/http-server.h"
-#include "sherpa/cpp_api/websocket/tee_stream.h"
+#include "sherpa/cpp_api/websocket/tee-stream.h"
 #include "websocketpp/config/asio_no_tls.hpp"  // TODO(fangjun): support TLS
 #include "websocketpp/server.hpp"
